1st Working version
* 1st working version that's fully compatible with the RelayServer.
Only tested running reconstruction using "python *reconscript.py*"
and "mpirun -n 1 python *reconscript.py*".
Iterations are currently not started until after end_of_scan == True.

* Closes the socket for recieving data when end_of_scan == True,
and tells the RelayServer that it's OK to close its corresponding
socket.

* load function now expects separate replies for the motor-messages
and the images. Images are expected to be compressed with lz4 when
recieved.

* Corrected the way that x- and y positions are extracted with xMotor
and yMotor from the recieved motor-message.